Fixed linking of libraries.
The linker searches libraries and object files in the order they are specified.
That means that the linker would process the libraries first before the object
files were processed, which caused it to not load the library functions for
the object files dependent on it.

Example:

‘foo.o -lz bar.o’ searches library ‘z’ after file foo.o but before bar.o. If
bar.o refers to functions in ‘z’, those functions may not be loaded.