Fixed a bug with setting sa_flags for sigaction
Since struct sigaction is allocated on the stack it will contain garbage values.
sa_flags kept these garbage values while setting SA_RESTART. Sometimes this
could cause it to have SA_RESETHAND set, which restores the signal action to the
default upon entry to the signal handler. In case of SIGALRM this is to
terminate the process.

Also modified the signal handler to just set a flag and then return.