Added support for NetBIOS name service protocol + a lot of bug fixes.
The char buffer I use to read the packets needs to be unsigned or else it will
be signed adjusted when I use it in calculations, e.g.:

uint16_t = buffer[0] << 8 | buffer[1]

If buffer[1] is greater than 2^7 - 1, it will be converted to "ffxx" instead of
"00xx" when used in 16 bits calculation. Implementing this made me discover some
bugs in the packet parsing code.