Implemented support for selecting a packet and printing more information
- For now this is only working for ARP but will start implementing support
for more protocols.
- Started to separate the different tasks more into own "modules", i.e.
packet.c is only used for dissecting packets and will not make any calls to
output.c to display anything in GUI. These modules will just present an API for
other parts to use. main.c will act as a controller (and for now also a model)
that ties the separate modules together.
- Added a dynamic array data structure (vector).