Fixed a bug in parse_dns_name
The length returned would be too great if the domain name had a pointer to a
sequence of labels ending in a pointer (which again can contain other pointers).
name_ptr_len should only be updated on the first pointer encountered.

This can be tested with e.g. "host d.dropbox.com".