Fixed some issues with getting input + some cleanup
TODO: Should make an own file for the main screen and let layout.c handle
general layout. The main screen is special in the sense that it's not part
of the screen stack, but is always present at the "bottom", i.e. how ncurses
handle stdscr. The status window should also not be part of the main screen.
Other screens can then make use of status to give feedback etc.