Fixed some issues with hexdump
For some reason the while loop printing the offsets could continue indefinitely
when incrementing the index as an argument to the mvwaddch instruction. This
would happen if the payload was too big to print on screen.

This still isn't working properly, as the remaining lines will be printed next
to each other at the bottom of the screen. Still better than everyting becoming
unresponsive, however.