Fixed some serious issues with starting/stopping a packet scan
The socket descriptor was not updated in the fds array, which caused poll to
return immediately when the descriptor was closed. This would make the main
loop enter a busy loop and use 100% of the cpu.

Also should not call run() again and push a new instance of it on the call
stack.