Enabled scrolling inside the pad
This is work in progress. The ui/scrolling code is a mess and needs to be
cleaned up. The way the scrolling is implemented is way too complex. I need
to rethink the design of this!

Need to change the names of subwindow, main_line, selected_line etc. to
something sensible so that one actually understands what they are used for.