Harmonized reading from file with performing a live capture
Both options will now use a callback that is called for each packet read.

Also removed read_packet from the decoder and moved the functionality to
initialize the socket and read from the network device to interface.c. This
will now be handled in a platform agnostic way. Still need to move a lot
of the platform dependent functionality to their respective directories.