bpf: Reject the packet if trying to access outside the packet bounds
The return value indicates how many bytes of the packet to accept (should be
usnigned really. Anyway the amount will be the minumum of the packet length and
the amount returned from the filter.)

Even with copiuos amount of beer this was an easy bug to find. Friday night
and all...