ui: Fixed a bug with "go to line"
"Go to line" is used to jump to the packet with the corresponding number.
When a filter is active the line number will not be the same as the packet
number and we therefore need to search for the correct packet. Renamed the
desciption to "go to packet" to better show the functionality of this.