ui: Fixed a bug when selecting protocol elements
When selecting headers/elements in a packet the view will be scrolled to
fit everything on screen. If the packet is too big it would always scroll
to the top even though the current element that is displayed is not able
to be shown.

Now it will stay where it is and only expand the current element.