ui: Fixed a bug with status window and following a TCP stream
When in "follow stream" mode and using 'g' to go to a specific packet and
then directly hitting escape, it would still be in "go to packet" mode. Instead
"follow stream" mode would be exited and it would return to main screen.

Also disabled filtering of packets when in "follow stream" mode.